+++ types/oolong.d.ts
@@ -14,6 +14,7 @@
 /// <reference path="std/encoding.d.ts" />
 /// <reference path="std/uuid.d.ts" />
 /// <reference path="std/semver.d.ts" />
+/// <reference path="std/fmt.d.ts" />
 /// <reference path="std/log.d.ts" />
 /// <reference path="std/http.d.ts" />
 /// <reference path="std/path.d.ts" />